Fix Entities

+++ src/main/java/WayofTime/alchemicalWizardry/common/entity/mob/EntityIceDemon.java
@@ -382,10 +382,10 @@ public class EntityIceDemon extends EntityDemon implements IRangedAttackMob
 
         if (par1)
         {
-            this.dataWatcher.updateObject(16, b0 | 2);
+            this.dataWatcher.updateObject(16, Byte.valueOf((byte)(b0 | 2)));
         } else
         {
-            this.dataWatcher.updateObject(16, b0 & -3);
+            this.dataWatcher.updateObject(16, Byte.valueOf((byte)(b0 & -3)));
         }
     }
